Improve tests by not only compiling but also running gem code
Previously the run_test.rb would only compile mruby with this gem and
see that that worked, but that's not really sufficient.

Now, it also executes mruby by itself to see that that works (this
caught an error I ran into where mruby 3 compiled with this gem but doesn't
run)

And then it also executes the example from the README which tests the
code paths in mrblib/require.rb.